+++ tests/manage-accounts.test.ts
@@ -1,16 +1,14 @@
+import test from 'tape'
 import { wasmGlobalsReady } from '@entropyxyz/sdk'
 // @ts-ignore
 import { isValidSubstrateAddress } from '@entropyxyz/sdk/utils'
 // @ts-ignore
 import Keyring from '@entropyxyz/sdk/keys'
 import { randomAsHex } from '@polkadot/util-crypto'
-import test from 'tape'
-// @ts-ignore
-import { spinNetworkUp, spinNetworkDown, } from "@entropyxyz/sdk/testing"
-import { EntropyAccountConfig, EntropyConfig } from 'src/config/types'
-import { listAccounts } from 'src/flows/manage-accounts/list'
-import { createAccount } from 'src/flows/manage-accounts/utils/create-account'
-import * as config from 'src/config'
+import { EntropyAccountConfig, EntropyConfig } from '../src/config/types'
+import { listAccounts } from '../src/flows/manage-accounts/list'
+import { createAccount } from '../src/flows/manage-accounts/utils/create-account'
+import * as config from '../src/config'
 import { promiseRunner, sleep } from './testing-utils'
 import { charlieStashAddress, charlieStashSeed } from './testing-utils/constants'
 
@@ -42,7 +40,7 @@ test('List Accounts', async t => {
   t.deepEqual(accountsArray, [{
     name: account.name,
     address: account.address,
-    verifyingKeys: account.data.admin.verifyingKeys
+    verifyingKeys: account?.data?.admin?.verifyingKeys
   }])
 
   // Resetting accounts on config to test for empty list
@@ -76,6 +74,6 @@ test('Create Account', async t => {
   const isValidAddress = isValidSubstrateAddress(newAccount.address)
 
   t.ok(isValidAddress, 'Valid address created')
-  t.equal(newAccount.address, admin.address, 'Generated Account matches Account created by Keyring')
+  t.equal(newAccount.address, admin?.address, 'Generated Account matches Account created by Keyring')
   t.end()
 })